+++ ydbsql/retry.go
@@ -1,6 +1,11 @@
 package ydbsql
 
-import "github.com/ydb-platform/ydb-go-sdk/v3"
+import (
+	"context"
+	"time"
+
+	"github.com/ydb-platform/ydb-go-sdk/v3"
+)
 
 type RetryConfig struct {
 	// MaxRetries is a number of maximum attempts to retry a failed operation.
@@ -11,10 +16,42 @@ type RetryConfig struct {
 	RetryChecker ydb.RetryChecker
 
 	// Backoff is a selected backoff policy.
-	// If backoff is nil, then the DefaultBackoff is used.
+	// Deprecated: use pair FastBackoff / SlowBackoff instead
 	Backoff ydb.Backoff
+
+	// FastBackoff is a selected backoff policy.
+	// If backoff is nil, then the ydb.DefaultFastBackoff is used.
+	FastBackoff ydb.Backoff
+
+	// SlowBackoff is a selected backoff policy.
+	// If backoff is nil, then the ydb.DefaultSlowBackoff is used.
+	SlowBackoff ydb.Backoff
+
+	// FastSlot is an init slot for fast retry
+	// If FastSlot is zero then the ydb.DefaultFastSlot is used.
+	FastSlot time.Duration
+
+	// SlowSlot is as zero then the ydb.DefaultSlowSlot is used.
+	SlowSlot time.Duration
 }
 
-func isBusy(err error) bool {
-	return retryChecker.Check(err).MustCheckSession()
+func backoff(ctx context.Context, m ydb.RetryMode, rc *RetryConfig, i int) error {
+	var b ydb.Backoff
+	switch m.BackoffType() {
+	case ydb.BackoffTypeNoBackoff:
+		return nil
+	case ydb.BackoffTypeFastBackoff:
+		if rc.FastBackoff != nil {
+			b = rc.FastBackoff
+		} else {
+			b = rc.Backoff
+		}
+	case ydb.BackoffTypeSlowBackoff:
+		if rc.SlowBackoff != nil {
+			b = rc.SlowBackoff
+		} else {
+			b = rc.Backoff
+		}
+	}
+	return ydb.WaitBackoff(ctx, b, i)
 }